pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "free-llm-router"
version = "0.1.0"
description = "OpenAI-like API router for permanent free-tier LLM providers"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn>=0.29,<1.0",
  "httpx>=0.27,<1.0",
  "pydantic>=1.10,<3.0",
  "PyYAML>=6.0",
  "streamlit>=1.32,<2.0",
]

[project.scripts]
free-llm-router = "free_llm_router.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
